# src/main/java/com/gila/ecommerce/controller/ProductController.java
package com.gila.ecommerce.controller;

import com.gila.ecommerce.api.ProductsApi;
import com.gila.ecommerce.dto.ImportProducts202Response;
import com.gila.ecommerce.dto.ProductDto;
import com.gila.ecommerce.dto.ProductImportStatusDto;
import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.RestController;
import org.springframework.web.multipart.MultipartFile;

import java.util.ArrayList;
import java.util.List;
import java.util.UUID;

/**
 * Controller stub implementing products REST catalog endpoints.
 */
@RestController
public class ProductController implements ProductsApi {

    /**
     * Create a new product.
     * @param productDto details of product to create
     * @return the created product details
     */
    @Override
    public ResponseEntity<ProductDto> createProduct(ProductDto productDto) {
        ProductDto created = new ProductDto();
        created.setId(UUID.randomUUID());
        created.setName(productDto.getName());
        created.setDescription(productDto.getDescription());
        created.setPrice(productDto.getPrice());
        created.setStock(productDto.getStock());
        created.setCategory(productDto.getCategory());
        return ResponseEntity.status(HttpStatus.CREATED).body(created);
    }

    /**
     * Delete an existing product by id.
     * @param id target product identifier
     * @return empty response entity with 204 status
     */
    @Override
    public ResponseEntity<Void> deleteProduct(UUID id) {
        return ResponseEntity.noContent().build();
    }

    /**
     * Retrieve status of a products CSV import request.
     * @param taskId CSV import task identifier
     * @return status report of the import process
     */
    @Override
    public ResponseEntity<ProductImportStatusDto> getImportStatus(UUID taskId) {
        ProductImportStatusDto status = new ProductImportStatusDto();
        status.setTaskId(taskId);
        status.setStatus("COMPLETED");
        status.setTotalRows(100);
        status.setProcessedRows(95);
        status.setErrorCount(5);
        status.setWarnings(new ArrayList<>());
        return ResponseEntity.ok(status);
    }

    /**
     * Retrieve details of a singular product by id.
     * @param id target product identifier
     * @return product details DTO
     */
    @Override
    public ResponseEntity<ProductDto> getProductById(UUID id) {
        ProductDto product = new ProductDto();
        product.setId(id);
        product.setName("Mock Product");
        product.setDescription("Mock Description");
        product.setPrice(9.99);
        product.setStock(50);
        product.setCategory("Books");
        return ResponseEntity.ok(product);
    }

    /**
     * Retrieve list of products with filters.
     * @param query search query
     * @param category category filter
     * @param page page number
     * @param size page limit size
     * @return list of matching products
     */
    @Override
    public ResponseEntity<List<ProductDto>> getProducts(String query, String category, Integer page, Integer size) {
        List<ProductDto> list = new ArrayList<>();
        ProductDto item = new ProductDto();
        item.setId(UUID.randomUUID());
        item.setName("Sample Item");
        item.setDescription("Sample Description");
        item.setPrice(19.99);
        item.setStock(10);
        item.setCategory("Electronics");
        list.add(item);
        return ResponseEntity.ok(list);
    }

    /**
     * Import products from a CSV file asynchronously.
     * @param file CSV file input stream resource
     * @return transaction reference taskId
     */
    @Override
    public ResponseEntity<ImportProducts202Response> importProducts(MultipartFile file) {
        ImportProducts202Response response = new ImportProducts202Response();
        response.setTaskId(UUID.randomUUID());
        response.setStatus("PENDING");
        return ResponseEntity.status(HttpStatus.ACCEPTED).body(response);
    }

    /**
     * Update details of an existing product.
     * @param id target product identifier
     * @param productDto updated product values
     * @return updated product details
     */
    @Override
    public ResponseEntity<ProductDto> updateProduct(UUID id, ProductDto productDto) {
        ProductDto updated = new ProductDto();
        updated.setId(id);
        updated.setName(productDto.getName());
        updated.setDescription(productDto.getDescription());
        updated.setPrice(productDto.getPrice());
        updated.setStock(productDto.getStock());
        updated.setCategory(productDto.getCategory());
        return ResponseEntity.ok(updated);
    }
}